requirements.txt
pytest-playwright==0.6.2
allure-pytest==2.13.5
docker==7.1.0